rename useMutationWithPermissionHandling and move it into its own file
