+++ src/api/broadcasts/mutate/updateBroadcasts.ts
@@ -7,7 +7,7 @@ import {
 import { apiFactory } from "~/api/apiFactory";
 import { broadcastKeys } from "..";
 import { useApiKey } from "~/util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useNamespace } from "~/util/store/namespace";
 import { useQueryClient } from "@tanstack/react-query";
 
@@ -40,7 +40,7 @@ export const useUpdateBroadcasts = ({
       },
     });
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn,
     onSuccess: (data) => {
       queryClient.setQueryData<BroadcastsResponseSchemaType>(
