+++ src/api/enterprise/groups/mutation/create.ts
@@ -3,7 +3,7 @@ import { GroupCreatedEditedSchema, GroupFormSchemaType } from "../schema";
 import { apiFactory } from "~/api/apiFactory";
 import { groupKeys } from "..";
 import { useApiKey } from "~/util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useNamespace } from "~/util/store/namespace";
 import { useQueryClient } from "@tanstack/react-query";
 import { useToast } from "~/design/Toast";
@@ -31,7 +31,7 @@ export const useCreateGroup = ({
     throw new Error("namespace is undefined");
   }
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn: (tokenFormProps: GroupFormSchemaType) =>
       createGroup({
         apiKey: apiKey ?? undefined,
