+++ src/api/enterprise/groups/mutation/edit.ts
@@ -3,7 +3,7 @@ import { GroupCreatedEditedSchema, GroupFormSchemaType } from "../schema";
 import { apiFactory } from "~/api/apiFactory";
 import { groupKeys } from "..";
 import { useApiKey } from "~/util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useNamespace } from "~/util/store/namespace";
 import { useQueryClient } from "@tanstack/react-query";
 import { useToast } from "~/design/Toast";
@@ -52,7 +52,7 @@ export const useEditGroup = ({
     throw new Error("namespace is undefined");
   }
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn: ({
       groupId,
       tokenFormProps,
