+++ src/api/enterprise/tokens/mutate/delete.ts
@@ -7,7 +7,7 @@ import {
 import { apiFactory } from "~/api/apiFactory";
 import { tokenKeys } from "..";
 import { useApiKey } from "~/util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useNamespace } from "~/util/store/namespace";
 import { useQueryClient } from "@tanstack/react-query";
 import { useToast } from "~/design/Toast";
@@ -58,7 +58,7 @@ export const useDeleteToken = ({
     throw new Error("namespace is undefined");
   }
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn: (token: TokenSchemaType) =>
       deleteToken({
         apiKey: apiKey ?? undefined,
