+++ src/api/errorHandling.ts
@@ -1,7 +1,3 @@
-import { UseMutationOptions, useMutation } from "@tanstack/react-query";
-
-import { t } from "i18next";
-import { useToast } from "~/design/Toast";
 import { z } from "zod";
 
 /**
@@ -69,41 +65,4 @@ export const getPermissionStatus = (error: unknown): PermissionStatus => {
   return {
     isAllowed: true,
   };
-};
-
-type UseMutationParam<TData, TError, TVariables> = UseMutationOptions<
-  TData,
-  TError,
-  TVariables
->;
-
-/**
- * useMutationWithPermissionHandling is a wrapper around useMutation that will
- * hook into the onError callback and check if the error is a permission error.
- * If it is, it will display a toast message to the user and early return. So
- * that no further error handling will be done.
- */
-export const useMutationWithPermissionHandling = <
-  TData = unknown,
-  TError = unknown,
-  TVariables = void
->(
-  useMutationParams: UseMutationParam<TData, TError, TVariables>
-) => {
-  const { toast } = useToast();
-  return useMutation({
-    ...useMutationParams,
-    onError: (error, variable, context) => {
-      const res = getPermissionStatus(error);
-      if (!res.isAllowed) {
-        toast({
-          title: t("api.generic.noPermissionTitle"),
-          description: res.message ?? t("api.generic.noPermissionDescription"),
-          variant: "error",
-        });
-        return;
-      }
-      useMutationParams.onError?.(error, variable, context);
-    },
-  });
 };
