+++ src/api/namespaces/mutate/createNamespace.ts
@@ -5,7 +5,7 @@ import { apiFactory } from "~/api/apiFactory";
 import { namespaceKeys } from "..";
 import { sortByName } from "~/api/tree/utils";
 import { useApiKey } from "~/util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useQueryClient } from "@tanstack/react-query";
 import { useToast } from "~/design/Toast";
 import { useTranslation } from "react-i18next";
@@ -26,7 +26,7 @@ export const useCreateNamespace = ({
   const queryClient = useQueryClient();
   const { t } = useTranslation();
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn: ({
       name,
       mirror,
