+++ src/api/namespaces/mutate/deleteNamespace.ts
@@ -2,7 +2,7 @@ import { NamespaceDeletedSchema } from "../schema";
 import { apiFactory } from "~/api/apiFactory";
 import { namespaceKeys } from "..";
 import { useApiKey } from "~/util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useQueryClient } from "@tanstack/react-query";
 import { useToast } from "~/design/Toast";
 import { useTranslation } from "react-i18next";
@@ -24,7 +24,7 @@ export const useDeleteNamespace = ({
   const queryClient = useQueryClient();
   const { t } = useTranslation();
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn: ({ namespace }: { namespace: string }) =>
       deleteNamespace({
         apiKey: apiKey ?? undefined,
