+++ src/api/registries/mutate/createRegistry.ts
@@ -7,7 +7,7 @@ import {
 import { apiFactory } from "~/api/apiFactory";
 import { registriesKeys } from "..";
 import { useApiKey } from "~/util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useNamespace } from "~/util/store/namespace";
 import { useQueryClient } from "@tanstack/react-query";
 import { useToast } from "~/design/Toast";
@@ -44,7 +44,7 @@ export const useCreateRegistry = ({
       },
     });
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn,
     onSuccess: (registry, variables) => {
       queryClient.invalidateQueries(
