+++ src/api/secrets/mutate/createSecret.ts
@@ -9,7 +9,7 @@ import { apiFactory } from "~/api/apiFactory";
 import { secretKeys } from "..";
 import { sortByName } from "~/api/tree/utils";
 import { useApiKey } from "~/util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useNamespace } from "~/util/store/namespace";
 import { useQueryClient } from "@tanstack/react-query";
 import { useToast } from "~/design/Toast";
@@ -59,7 +59,7 @@ export const useCreateSecret = ({
     throw new Error("namespace is undefined");
   }
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn: ({ name, value }: { name: string; value: string }) =>
       createSecret({
         apiKey: apiKey ?? undefined,
