+++ src/api/secrets/mutate/deleteSecret.ts
@@ -7,7 +7,7 @@ import {
 import { apiFactory } from "../../apiFactory";
 import { secretKeys } from "..";
 import { useApiKey } from "../../../util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useNamespace } from "../../../util/store/namespace";
 import { useQueryClient } from "@tanstack/react-query";
 import { useToast } from "../../../design/Toast";
@@ -54,7 +54,7 @@ export const useDeleteSecret = ({
     throw new Error("namespace is undefined");
   }
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn: ({ secret }: { secret: SecretSchemaType }) =>
       deleteSecret({
         apiKey: apiKey ?? undefined,
