+++ src/api/services/mutate/deleteRevision.ts
@@ -6,7 +6,7 @@ import {
 import { apiFactory } from "~/api/apiFactory";
 import { serviceKeys } from "..";
 import { useApiKey } from "~/util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useNamespace } from "~/util/store/namespace";
 import { useQueryClient } from "@tanstack/react-query";
 import { useToast } from "~/design/Toast";
@@ -56,7 +56,7 @@ export const useDeleteServiceRevision = ({
     throw new Error("namespace is undefined");
   }
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn: ({
       service,
       revision,
