+++ src/api/variables/mutate/deleteVariable.ts
@@ -2,7 +2,7 @@ import { VarDeletedSchema, VarSchemaType } from "../schema";
 
 import { apiFactory } from "~/api/apiFactory";
 import { useApiKey } from "~/util/store/apiKey";
-import { useMutationWithPermissionHandling } from "~/api/errorHandling";
+import useMutationWithPermissions from "~/api/useMutationWithPermissions";
 import { useNamespace } from "~/util/store/namespace";
 import { useQueryClient } from "@tanstack/react-query";
 import { useToast } from "~/design/Toast";
@@ -40,7 +40,7 @@ export const useDeleteVar = ({
       },
     });
 
-  return useMutationWithPermissionHandling({
+  return useMutationWithPermissions({
     mutationFn,
     onSuccess: (_, variables) => {
       queryClient.invalidateQueries(
